Got rid of our own ndarray subclass.
This had some advantages (e.g. allowed A.dot(B) syntax) but
plenty of drawbacks: effort to check output types and recast
as necessary, different behavior of keepdims and scalar-valued
outputs, and wrapping arrays coming from other libraries
(e.g. scipy optimize). We may reintroduce it if we need it
in order to implement assignment.