vspace ones and standard_basis methods (#236)
* add vspace.ones() and .standard_basis(), used in grad and jacobian.

the main advantage is that it reduces the explicit dependence on numpy, so that
new numerical libraries that register their own vspaces can also build the
inputs to the vector-jacobian product chain in grad and jacobian. the code is
also cleaner.

* remove grad calls to scalar_fun and as_scalar, comment out tests for error on
evaluating a grad on a non-scalar-output function

this changes the behavior of grad: instead of raising an error when the
function output is non-scalar, grad acts like elementwise_grad, which is
usually the correct behavior. we could add the error check back to grad by
checking the output vspace.

* remove unused function get_vspace