Adjust make_vjp and backward_pass to incrementally free nodes when possible.
Also introduce the optional argument "preserve_tape" to make_vjp, which enables
this optimization when we can guarantee the vjp function is evaluated only
once (such as in `grad`).

I haven't tested that the ref counts are actually going to zero as intended, so
this is really a draft commit. However, the vjp function no longer closes over
end_node (or start_node), and the tape now gets consumed incrementally, so I
think this is at least the right idea.