Clarify that elementwise_grad is only applicable to diagonal Jacobians
I think it's a limitation that is important to point out. Trying to use elementwise_grad for non-diagonal Jacobians will silently return incorrect values. 

I've copied and modified the doc from some older version of elementwise_grad. The doc could be extended further if there are any other features/limitations to be pointed out for the current version.