wrap_namespace maps multiple names of the same object to the same wrapped object
Avoids creating two different wrappers for the same object.

Generally cleaner, but more importantly towards better ufuncs support,
which will require mapping from ufuncs to their wrapped version.

Another minor pro is that this theoretically avoids needing to define grads
for two names of the same object, e.g., we could define the VJP for anp.conj
but not for anp.conjugate since they are actually the same wrapped object.
However, it's minor and theoretical because whether or not two "equivalent"
numpy functions/ufuncs are the same object is apparently an implementation
detail, e.g., np.amax is documented as an alias of np.max, but they actually
aren't the same object. Since this can probably change between numpy versions,
the robust thing to do would to keep the grad definitions for both names.